Adicionado o calculo de area e perimetro do quadrado

+++ Nivelamento/src/algoritmo/Distancia.java
@@ -3,7 +3,7 @@ package algoritmo;
 public class Distancia {
 
     public static double coordenada(int Xa, int Xb, int Ya, int Yb){
-        int resultado = (int) Math.sqrt(Math.pow(Xb - Xa, 2) + Math.pow(Yb - Ya, 2));
+        double resultado = (double) Math.sqrt(Math.pow(Xb - Xa, 2) + Math.pow(Yb - Ya, 2));
         return resultado;
     }
 }
